objBboxFn_800640cc: 11th param is u8 not int (retail reads lbz r22,235(r1), we read lwz r31,232(r1) — same slot, low byte, big-endian)
Correctness fix proven from the retail prologue: frame is 224, so the first
stack argument lives at 232(r1) and retail loads it byte-wide at 235. The
prologue is now instruction-identical to retail modulo register numbers.

Fuzzy trades down (fn 96.95169 -> 96.27536, unit 98.56370 -> 98.53751): the
narrower type flips the whole saved-register bank between two bands (params
r24-r30/arg8 r31 before, params r15-r21 after; retail is params r25-r31/arg8
r22). Declaration-order sweeps at both scopes are bit-inert, so the residual
is register allocation, not source shape.

Tree-wide sweep of this defect class: 1 hit in 7843 paired functions.

+++ include/main/track_bbox_api.h
@@ -39,6 +39,6 @@ typedef struct TrackBBoxHit
 STATIC_ASSERT(sizeof(TrackBBoxHit) == 0x54);
 
 int objBboxFn_800640cc(f32* from, f32* to, f32 radius, int mode, TrackBBoxHit* hit, GameObject* self, int flags,
-                       int mask, int slot, int arg10);
+                       int mask, int slot, u8 arg10);
 
 #endif /* MAIN_TRACK_BBOX_API_H_ */
